requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fpdf==1.7.2
fsspec==2025.3.2
gunicorn==23.0.0
imageio==2.37.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==11.1.0
ptflops==0.7.4
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.3.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
Werkzeug==3.1.3
